Move integration test utils to util package
